+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-### v?.?.? / ????.??.??
+### v0.5.0 / ????.??.??
 
+- feat: Studio Pack Generator GUI !
 - fix: exit if nightMode && (autoNextStoryTransition || selectNextStoryAtEnd)
 - fix #25: metadata.json "nightmode" is not extracted correctly by @lasconic and
   @jersou in #26